pyproject.toml
[tool.poetry]
name = "draftdiff"
version = "0.1.7"
description = "creating dota2 counterpicker"
authors = ["Jeff", "Luke"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "draftdiff", from = "." }
]

[tool.poetry.dependencies]
python = "3.11.0"
beautifulsoup4 = "^4.12.3"
pandas = "^2.2.1"
loguru = "^0.7.2"
tqdm = "^4.66.2"
requests = "^2.31.0"
google-api-python-client = "^2.125.0"
oauth2client = "^4.1.3"
google-auth = "^2.29.0"
gspread = "^6.1.0"
boto3 = "^1.34.122"
awscli = "^1.33.4"
pyarrow = "^16.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
select = [
    "I"
]
src = ["draftdiff"]

[tool.black]
line-length = 75
target-version = ['py311']
include = '\.pyi?$'
exclude = '''

/(
    \.eggs         # exclude a few common directories in the
  | \.git          # root of the project
  | \.hg
  | \.mypy_cache
  | \.nox
  | \.tox
  | \.venv
  | _build
  | build
  | dist
)/

'''